Fixed group.dn and u.dn to distinguished name - was undefined before
This is the PR for this issue #84 

I really don't know why this fixed the problem but I saw in the debugger that `group.dn` and `u.dn` was undefined. After I changed it to the correct attribute the found user actually has on the object `distinguishedName` it's working now. I will receive now all the users from my nested groups that have other groups nested inside.